split node-ffi
created node-ffi-buffer.d.ts partial

- ref.d.ts
- ref-array.d.ts
- ref-struct.d.ts
- ref-union.d.ts
